Refactor (#42): refactoring based on PR feedback
