+++ frontend/src/routes/ArticlePage/articlePagestyles.ts
@@ -1,0 +1,8 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export const articlePageStyles = makeStyles(theme => ({
+    container: {
+        width: "80%",
+        margin: "auto",
+    },
+}))
