+++ frontend/src/routes/HomePage/homePageStyles.ts
@@ -1,0 +1,8 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export const homePageStyles = makeStyles(theme => ({
+    container: {
+        width: "80%",
+        margin: "auto",
+    },
+}))
