+++ frontend/src/routes/NewsPage/newsPageStyles.ts
@@ -1,0 +1,8 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export const newsPageStyles = makeStyles(theme => ({
+    container: {
+        width: "80%",
+        margin: "auto",
+    },
+}))
